separated waybill layout for reusability
